+++ src/controllers/contact.php
@@ -1,0 +1,11 @@
+<?php
+
+    //require_once("src/model/contact.php");
+
+    function contact()
+    {
+        //$data = generate();
+        // $data = getProjects();
+
+        require("template/pages/contact.php");
+    }