+++ src/controllers/cv.php
@@ -1,0 +1,11 @@
+<?php
+
+    //require_once("src/model/cv.php");
+
+    function cv()
+    {
+        //$data = generate();
+        // $data = getProjects();
+
+        require("template/pages/cv.php");
+    }