+++ src/controllers/home.php
@@ -1,0 +1,11 @@
+<?php
+
+    //require_once("src/model/home.php");
+
+    function home()
+    {
+        //$data = generate();
+        // $data = getProjects();
+
+        require("template/pages/home.php");
+    }