+++ src/model/database/secret.php
@@ -1,0 +1,5 @@
+<?php
+    $host='localhost';
+    $user='root';
+    $userpswd='';
+    $db='portfolio_db_project';