+++ template/layout.php
@@ -13,6 +13,7 @@
     <body>
         <nav>
             <ul>
+                <li><a href="index.php?link=home">Accueil</a></li>
                 <li><a href="./index.php?link=cv">CV</a></li>
                 <li><a href="./index.php?link=projectList">Projet</a></li>
                 <li><a href="./index.php?link=contact">Contact</a></li>