+++ template/pages/contact.php
@@ -1,0 +1,5 @@
+<?php $title = "Contact"; ?>
+
+<?php $content = "page non disponible"; ?>
+
+<?php require('template/layout.php'); ?>