+++ template/pages/cv.php
@@ -1,0 +1,5 @@
+<?php $title = "cv"; ?>
+
+<?php $content = "page non disponible"; ?>
+
+<?php require('template/layout.php'); ?>