+++ template/pages/home.php
@@ -1,0 +1,5 @@
+<?php $title = "accueil"; ?>
+
+<?php $content = "page non disponible"; ?>
+
+<?php require('template/layout.php'); ?>